fix(fixtures): Fix naming
Switch from questionparser's parser to get_questions and from loader.py
to main.py
